# bench(http): optimize to_frames and to_lines functions